+++ spec/samsao_config_spec.rb
@@ -1,0 +1,30 @@
+require File.expand_path('../spec_helper', __FILE__)
+
+module Danger
+  describe Danger::DangerSamsao do
+    describe 'with Dangerfile' do
+      before do
+        @dangerfile = testing_dangerfile
+        @plugin = @dangerfile.samsao
+      end
+
+      describe 'samsao_config' do
+        it 'can configure single source' do
+          @plugin.config do
+            sources 'app/src'
+          end
+
+          expect(@plugin.config.sources).to eq(['app/src'])
+        end
+
+        it 'can configure multiple sources' do
+          @plugin.config do
+            sources 'app/src', 'lib/src'
+          end
+
+          expect(@plugin.config.sources).to eq(['app/src', 'lib/src'])
+        end
+      end
+    end
+  end
+end