+++ spec/samsao_feature_single_commit_spec.rb
@@ -1,0 +1,42 @@
+require File.expand_path('../spec_helper', __FILE__)
+
+module Danger
+  describe Danger::DangerSamsao do
+    describe 'with Dangerfile' do
+      before do
+        @dangerfile = testing_dangerfile
+        @plugin = @dangerfile.samsao
+      end
+
+      describe 'feature branch commits' do
+        it 'continues on feature branch and one commit' do
+          allow(@plugin.github).to receive(:branch_for_head).and_return('feature/a')
+          allow(@plugin.git).to receive(:commits).and_return(['sha1'])
+
+          @plugin.fail_when_non_single_commit_feature
+
+          expect(@dangerfile).to have_no_error
+        end
+
+        it 'continues on non-feature branch and multiple commits' do
+          allow(@plugin.github).to receive(:branch_for_head).and_return('fix/a')
+          allow(@plugin.git).to receive(:commits).and_return(['sha1', 'sha2'])
+
+          @plugin.fail_when_non_single_commit_feature
+
+          expect(@dangerfile).to have_no_error
+        end
+
+        it 'fails on feature branch and multiple commits' do
+          allow(@plugin.github).to receive(:branch_for_head).and_return('feature/a')
+          allow(@plugin.git).to receive(:commits).and_return(['sha1', 'sha2'])
+
+          @plugin.fail_when_non_single_commit_feature
+
+          expected_message = 'Your feature branch should have a single commit but found 2, squash them together!'
+          expect(@dangerfile).to have_error(expected_message)
+        end
+      end
+    end
+  end
+end