+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## In progress
 
+ * Added branch helpers actions:
+    * `fix_branch?`
+    * `feature_branch?`
+    * `release_branch?`
+    * `trivial_branch?`
+
  * Added `fail_when_changelog_update_missing` action.
 
  * Added `fail_when_non_single_commit_feature` action.